EPMLSTRTAI-374: modified removeCases method to require only list of suit ID and case IDs and written tests for this method

+++ src/main/java/com/epam/test_generator/controllers/CaseController.java
@@ -5,15 +5,24 @@ import com.epam.test_generator.dto.SuitDTO;
 import com.epam.test_generator.dto.ValidationErrorsDTO;
 import com.epam.test_generator.services.CaseService;
 import com.epam.test_generator.services.SuitService;
-import io.swagger.annotations.*;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-import java.util.List;
-import java.util.stream.Collectors;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class CaseController {
@@ -114,16 +123,26 @@ public class CaseController {
     @ApiOperation(value = "Delete cases from suit", nickname = "removeCases")
     @ApiResponses(value = {
         @ApiResponse(code = 200, message = "OK"),
-        @ApiResponse(code = 404, message = "Suit not found")
+        @ApiResponse(code = 404, message = "Suit not found"),
+        @ApiResponse(code = 400, message = "Suit doesn't contains some of cases")
     })
     @ApiImplicitParams({
-        @ApiImplicitParam(name = "suitId", value = "ID of suit which contains the case", required = true, dataType = "long", paramType = "path"),
-        @ApiImplicitParam(name = "suitDTO", value = "Suit object with casesIds", required = true, dataType = "SuitDTO", paramType = "body")
+        @ApiImplicitParam(name = "suitId", value = "ID of suit which contains the case",
+            required = true, dataType = "long", paramType = "path"),
+        @ApiImplicitParam(name = "removeCaseIds", value = "IDs of cases to be removed",
+            required = true, dataType = "Long[]", paramType = "body")
     })
     @RequestMapping(value = "/suits/{suitId}/cases", method = RequestMethod.DELETE, consumes = "application/json")
     public ResponseEntity<Void> removeCases(@PathVariable("suitId") long suitId,
-                                            @RequestBody @Valid SuitDTO suitDTO) {
+                                            @RequestBody Long[] removeCaseIds) {
+        SuitDTO suitDTO = suitService.getSuit(suitId);
+        if (!suitDTO.getCases().stream().map(CaseDTO::getId).collect(Collectors.toList())
+            .containsAll(
+                Arrays.asList(removeCaseIds))) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         List<Long> caseIds = suitDTO.getCases().stream().map(CaseDTO::getId)
+            .filter(Arrays.asList(removeCaseIds)::contains)
             .collect(Collectors.toList());
         caseService.removeCases(suitId, caseIds);
